requirements.txt
openai>=0.27.0
anthropic>=0.3.0
python-dotenv>=0.19.0
rich>=10.0.0
requests>=2.25.0